# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Openlocal API
  version: 0.1.0
  description: API for Openlocal local commerce inventory, storefronts, marketplace discovery, and analytics.
servers:
  - url: http://localhost:8080
...
security:
  - clerkBearer: []
components:
  securitySchemes:
    clerkBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      required:
        - error
...
    BusinessRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 160
          example: Casa Pan Local
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          example: casa-pan-local
        description:
          type: string
          example: Small-batch sourdough, pastries, and pantry goods baked every morning.
        business_type:
          type: string
          example: bakery
        phone:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
        logo_url:
          type: string
          format: uri
          nullable: true
        cover_image_url:
          type: string
          format: uri
          nullable: true
        status:
          type: string
          enum: [draft, active, suspended, archived]
          default: draft
        address:
          type: string
          nullable: true
        neighborhood:
          type: string
          nullable: true
          example: Condesa
        city:
          type: string
          default: CDMX
        state:
          type: string
          default: CDMX
        country:
          type: string
          default: MX
        postal_code:
          type: string
          nullable: true
        latitude:
          type: string
          nullable: true
          example: "19.411700"
        longitude:
          type: string
          nullable: true
          example: "-99.171800"
        pickup_available:
          type: boolean
          default: true
        delivery_available:
          type: boolean
          default: false
        timezone:
          type: string
          minLength: 1
          maxLength: 100
          default: America/Mexico_City
          example: America/Mexico_City
          description: IANA timezone used to interpret the business opening hours.
      required:
        - name
        - description
    Business:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        clerk_org_id:
          type: string
          description: Only returned on authenticated private business responses.
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        business_type:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
          format: email
        website:
          type: string
          format: uri
        logo_url:
          type: string
          format: uri
        cover_image_url:
          type: string
          format: uri
        status:
          type: string
        address:
          type: string
        neighborhood:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postal_code:
          type: string
        latitude:
          type: string
        longitude:
          type: string
        pickup_available:
          type: boolean
        delivery_available:
          type: boolean
        timezone:
          type: string
          example: America/Mexico_City
          description: IANA timezone used to interpret the business opening hours.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - description
        - business_type
        - city
        - state
        - country
        - pickup_available
        - delivery_available
        - timezone
    BusinessHourRequest:
      type: object
      additionalProperties: false
      properties:
        day_of_week:
          type: integer
          format: int32
          minimum: 0
          maximum: 6
          description: Day of the week, where 0 is Sunday and 6 is Saturday.
          example: 1
        opens_at:
          type: string
          nullable: true
          pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$"
          example: "09:00"
          description: Required in HH:MM format when the business is open; omit for closed days.
        closes_at:
          type: string
          nullable: true
          pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$"
          example: "18:00"
          description: Required in HH:MM format when the business is open; omit for closed days. May be earlier than opens_at for overnight hours.
        is_closed:
          type: boolean
          default: false
      required:
        - day_of_week
    BusinessHoursRequest:
      type: object
      additionalProperties: false
      properties:
        hours:
          type: array
          maxItems: 7
          uniqueItems: true
          description: Complete replacement schedule. Send an empty array to remove all configured hours; each day may appear at most once.
          items:
            $ref: "#/components/schemas/BusinessHourRequest"
      required:
        - hours
    BusinessHour:
      type: object
      additionalProperties: false
      properties:
        day_of_week:
          type: integer
          format: int32
          minimum: 0
          maximum: 6
          description: Day of the week, where 0 is Sunday and 6 is Saturday.
        opens_at:
          type: string
          nullable: true
          pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$"
        closes_at:
          type: string
          nullable: true
          pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$"
        is_closed:
          type: boolean
      required:
        - day_of_week
        - is_closed
    BusinessHours:
      type: object
      additionalProperties: false
      properties:
        timezone:
          type: string
          example: America/Mexico_City
          description: IANA timezone used to interpret every time in the schedule.
        hours:
          type: array
          items:
            $ref: "#/components/schemas/BusinessHour"
      required:
        - timezone
        - hours
...
paths:
...
  /api/v1/businesses:
    post:
      tags:
        - Businesses
      description: Creates a Clerk organization and the matching Openlocal business.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessRequest"
      responses:
        "201":
          description: Business created. Response includes clerk_org_id for frontend org activation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "502":
          description: Clerk organization creation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/businesses/me:
    get:
      tags:
        - Businesses
      description: Resolves the current Openlocal business from the active Clerk organization claim.
      responses:
        "200":
          description: Private business profile for the active Clerk organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "403":
          description: Active Clerk organization or local business membership is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No Openlocal business is linked to the active Clerk organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/businesses/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Businesses
      responses:
        "200":
          description: Private business profile for an authorized member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    patch:
      tags:
        - Businesses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessRequest"
      responses:
        "200":
          description: Updated private business profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /api/v1/businesses/{id}/hours:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Businesses
      description: Returns the opening-hours schedule for an authorized business member.
      responses:
        "200":
          description: Business opening hours ordered by day of week.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessHours"
        "403":
          description: Owner, manager, or staff membership is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
        - Businesses
      description: Atomically replaces the complete opening-hours schedule. Overnight ranges are supported.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessHoursRequest"
      responses:
        "200":
          description: Updated business opening hours ordered by day of week.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessHours"
        "400":
          description: Invalid weekday, duplicate day, or invalid opening-hours combination.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Owner or manager membership is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...