# internal/platform/postgres/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"database/sql"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
	"github.com/shopspring/decimal"
)

type AuditLog struct {
	ID          uuid.UUID          `json:"id"`
	BusinessID  uuid.NullUUID      `json:"business_id"`
	ActorUserID uuid.NullUUID      `json:"actor_user_id"`
	Action      string             `json:"action"`
	EntityType  string             `json:"entity_type"`
	EntityID    uuid.NullUUID      `json:"entity_id"`
	Metadata    []byte             `json:"metadata"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type Business struct {
	ID                uuid.UUID           `json:"id"`
	Name              string              `json:"name"`
	Slug              string              `json:"slug"`
	Description       string              `json:"description"`
	BusinessType      string              `json:"business_type"`
	Phone             sql.NullString      `json:"phone"`
	Whatsapp          sql.NullString      `json:"whatsapp"`
	Email             sql.NullString      `json:"email"`
	Website           sql.NullString      `json:"website"`
	LogoUrl           sql.NullString      `json:"logo_url"`
	CoverImageUrl     sql.NullString      `json:"cover_image_url"`
	Status            string              `json:"status"`
	Address           sql.NullString      `json:"address"`
	Neighborhood      sql.NullString      `json:"neighborhood"`
	City              string              `json:"city"`
	State             string              `json:"state"`
	Country           string              `json:"country"`
	PostalCode        sql.NullString      `json:"postal_code"`
	Latitude          decimal.NullDecimal `json:"latitude"`
	Longitude         decimal.NullDecimal `json:"longitude"`
	PickupAvailable   bool                `json:"pickup_available"`
	DeliveryAvailable bool                `json:"delivery_available"`
	CreatedAt         pgtype.Timestamptz  `json:"created_at"`
	UpdatedAt         pgtype.Timestamptz  `json:"updated_at"`
	ClerkOrgID        sql.NullString      `json:"clerk_org_id"`
	Timezone          string              `json:"timezone"`
}

type BusinessCategory struct {
	ID         uuid.UUID `json:"id"`
	BusinessID uuid.UUID `json:"business_id"`
	Name       string    `json:"name"`
	Slug       string    `json:"slug"`
}

type BusinessHour struct {
	ID         uuid.UUID   `json:"id"`
	BusinessID uuid.UUID   `json:"business_id"`
	DayOfWeek  int32       `json:"day_of_week"`
	OpensAt    pgtype.Time `json:"opens_at"`
	ClosesAt   pgtype.Time `json:"closes_at"`
	IsClosed   bool        `json:"is_closed"`
}

type BusinessMember struct {
	ID         uuid.UUID          `json:"id"`
	BusinessID uuid.UUID          `json:"business_id"`
	UserID     uuid.UUID          `json:"user_id"`
	ClerkOrgID string             `json:"clerk_org_id"`
	Role       string             `json:"role"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	UpdatedAt  pgtype.Timestamptz `json:"updated_at"`
}

type Category struct {
	ID         uuid.UUID          `json:"id"`
	BusinessID uuid.UUID          `json:"business_id"`
	Name       string             `json:"name"`
	Slug       string             `json:"slug"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type InventoryLocation struct {
	ID         uuid.UUID          `json:"id"`
	BusinessID uuid.UUID          `json:"business_id"`
	Name       string             `json:"name"`
	IsDefault  bool               `json:"is_default"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type Product struct {
	ID          uuid.UUID          `json:"id"`
	BusinessID  uuid.UUID          `json:"business_id"`
	CategoryID  uuid.NullUUID      `json:"category_id"`
	Name        string             `json:"name"`
	Slug        string             `json:"slug"`
	Description string             `json:"description"`
	Brand       sql.NullString     `json:"brand"`
	Unit        string             `json:"unit"`
	ProductType string             `json:"product_type"`
	IsHandmade  bool               `json:"is_handmade"`
	IsPublic    bool               `json:"is_public"`
	Status      string             `json:"status"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}

type ProductImage struct {
	ID        uuid.UUID          `json:"id"`
	ProductID uuid.UUID          `json:"product_id"`
	Url       string             `json:"url"`
	AltText   string             `json:"alt_text"`
	Position  int32              `json:"position"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type ProductSalesSummary struct {
	BusinessID   uuid.UUID       `json:"business_id"`
	VariantID    uuid.UUID       `json:"variant_id"`
	ProductID    uuid.UUID       `json:"product_id"`
	UnitsSold    decimal.Decimal `json:"units_sold"`
	SalesRevenue decimal.Decimal `json:"sales_revenue"`
}

type ProductTag struct {
	ProductID uuid.UUID `json:"product_id"`
	TagID     uuid.UUID `json:"tag_id"`
}

type ProductVariant struct {
	ID                uuid.UUID           `json:"id"`
	ProductID         uuid.UUID           `json:"product_id"`
	BusinessID        uuid.UUID           `json:"business_id"`
	Sku               string              `json:"sku"`
	Barcode           sql.NullString      `json:"barcode"`
	InternalCode      string              `json:"internal_code"`
	Name              string              `json:"name"`
	Attributes        []byte              `json:"attributes"`
	Price             decimal.Decimal     `json:"price"`
	Cost              decimal.NullDecimal `json:"cost"`
	Currency          string              `json:"currency"`
	TrackInventory    bool                `json:"track_inventory"`
	PublicStockStatus string              `json:"public_stock_status"`
	ReorderPoint      decimal.Decimal     `json:"reorder_point"`
	LeadTimeDays      int32               `json:"lead_time_days"`
	Status            string              `json:"status"`
	CreatedAt         pgtype.Timestamptz  `json:"created_at"`
	UpdatedAt         pgtype.Timestamptz  `json:"updated_at"`
}

type StockLevel struct {
	BusinessID       uuid.UUID          `json:"business_id"`
	VariantID        uuid.UUID          `json:"variant_id"`
	LocationID       uuid.UUID          `json:"location_id"`
	QuantityOnHand   decimal.Decimal    `json:"quantity_on_hand"`
	QuantityReserved decimal.Decimal    `json:"quantity_reserved"`
	UpdatedAt        pgtype.Timestamptz `json:"updated_at"`
}

type StockMovement struct {
	ID             uuid.UUID           `json:"id"`
	BusinessID     uuid.UUID           `json:"business_id"`
	VariantID      uuid.UUID           `json:"variant_id"`
	LocationID     uuid.UUID           `json:"location_id"`
	MovementType   string              `json:"movement_type"`
	Quantity       decimal.Decimal     `json:"quantity"`
	UnitCost       decimal.NullDecimal `json:"unit_cost"`
	ReferenceType  sql.NullString      `json:"reference_type"`
	ReferenceID    sql.NullString      `json:"reference_id"`
	Notes          string              `json:"notes"`
	CreatedBy      uuid.NullUUID       `json:"created_by"`
	CreatedAt      pgtype.Timestamptz  `json:"created_at"`
	IdempotencyKey sql.NullString      `json:"idempotency_key"`
}

type Tag struct {
	ID         uuid.UUID `json:"id"`
	BusinessID uuid.UUID `json:"business_id"`
	Name       string    `json:"name"`
	Slug       string    `json:"slug"`
}

type User struct {
	ID          uuid.UUID          `json:"id"`
	ClerkUserID string             `json:"clerk_user_id"`
	Email       sql.NullString     `json:"email"`
	FirstName   sql.NullString     `json:"first_name"`
	LastName    sql.NullString     `json:"last_name"`
	ImageUrl    sql.NullString     `json:"image_url"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}